core: Inherit the atomics module
